requirements.txt
# ==============================
# IMAGE PROCESSING DASHBOARD
# Dependencies for Edge Detection App
# ==============================

# Core image processing
numpy
pillow

# Data and math utilities
pandas

# Web framework (choose one depending on deployment)
streamlit

#python -m venv venv
#venv\Scripts\activate